# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: PipeForge API
  version: '1.0'
  description: >
    REST API for PipeForge — a visual data-processing pipeline builder.
    Browser clients authenticate via an httpOnly cookie set on login/register;
    non-browser clients may instead send `Authorization: Bearer <token>`.
servers:
  - url: /api
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FileSummary:
      type: object
      properties:
        id: { type: string }
        filePath: { type: string }
        originalName: { type: string }
        size: { type: integer }
        createdAt: { type: string, format: date-time }
        project:
          type: object
          description: 'Only present on the cross-project GET /files aggregate, not on the project-nested GET .../files.'
          properties: { id: { type: string, nullable: true }, name: { type: string, nullable: true } }
...
    Error:
      type: object
      properties:
        error: { type: string }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /projects/{projectId}/files/upload:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string } }
    post:
      summary: Upload a .csv or .json dataset to this project (used as a pipeline's csv-input/json-input source), shared with every project member
      description: Requires editor+ on the project. Storage quota is charged to the uploader personally, not the project.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  filePath: { type: string, description: 'Pass this as a csv-input/json-input node''s filePath config' }
                  originalName: { type: string }
                  size: { type: integer }
        '400': { description: 'No file, wrong type, or over the per-file size limit' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { description: "Would exceed the uploader's total storage quota" }
  /projects/{projectId}/files:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string } }
    get:
      summary: List this project's uploaded datasets (viewer+)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FileSummary' }
        '404': { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/files/{fileId}:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string } }
      - { name: fileId, in: path, required: true, schema: { type: string } }
    delete:
      summary: Delete an uploaded dataset (editor+; frees the uploader's storage quota)
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...